fix(engine): separate compensation subscription behaviour from other behaviour
Created a new behaviour that handle all the methods of the new
compensation subscription, this new behaviour should be used inside the
processors
