+++ engine/src/main/java/io/camunda/zeebe/engine/processing/bpmn/behavior/BpmnBehaviors.java
@@ -53,4 +53,6 @@ public interface BpmnBehaviors {
   BpmnJobActivationBehavior jobActivationBehavior();
 
   BpmnUserTaskBehavior userTaskBehavior();
+
+  BpmnCompensationSubscriptionBehaviour compensationSubscriptionBehaviour();
 }
