+++ engine/src/main/java/io/camunda/zeebe/engine/processing/bpmn/event/EndEventProcessor.java
@@ -12,6 +12,7 @@ import static io.camunda.zeebe.util.EnsureUtil.ensureNotNull;
 import io.camunda.zeebe.engine.processing.bpmn.BpmnElementContext;
 import io.camunda.zeebe.engine.processing.bpmn.BpmnElementProcessor;
 import io.camunda.zeebe.engine.processing.bpmn.behavior.BpmnBehaviors;
+import io.camunda.zeebe.engine.processing.bpmn.behavior.BpmnCompensationSubscriptionBehaviour;
 import io.camunda.zeebe.engine.processing.bpmn.behavior.BpmnEventPublicationBehavior;
 import io.camunda.zeebe.engine.processing.bpmn.behavior.BpmnIncidentBehavior;
 import io.camunda.zeebe.engine.processing.bpmn.behavior.BpmnJobBehavior;
@@ -46,6 +47,7 @@ public final class EndEventProcessor implements BpmnElementProcessor<ExecutableE
   private final BpmnJobBehavior jobBehavior;
   private final BpmnSignalBehavior signalBehavior;
   private final BpmnStateBehavior stateBehavior;
+  private final BpmnCompensationSubscriptionBehaviour compensationSubscriptionBehaviour;
 
   public EndEventProcessor(
       final BpmnBehaviors bpmnBehaviors,
@@ -58,6 +60,7 @@ public final class EndEventProcessor implements BpmnElementProcessor<ExecutableE
     jobBehavior = bpmnBehaviors.jobBehavior();
     stateBehavior = bpmnBehaviors.stateBehavior();
     signalBehavior = bpmnBehaviors.signalBehavior();
+    compensationSubscriptionBehaviour = bpmnBehaviors.compensationSubscriptionBehaviour();
   }
 
   @Override
@@ -325,11 +328,12 @@ public final class EndEventProcessor implements BpmnElementProcessor<ExecutableE
 
     @Override
     public void onActivate(final ExecutableEndEvent element, final BpmnElementContext activating) {
+      compensationSubscriptionBehaviour.updateCompensationSubscription(activating);
       final var activated =
           stateTransitionBehavior.transitionToActivated(activating, element.getEventType());
       // check for activities that are completed and have compensation handlers
       final Set<String> completedActivities =
-          stateBehavior.getCompletedActivitiesToCompensate(activated);
+          compensationSubscriptionBehaviour.getCompletedActivitiesToCompensate(activated);
 
       if (completedActivities.isEmpty()) {
         final var completing = stateTransitionBehavior.transitionToCompleting(activated);
@@ -338,7 +342,7 @@ public final class EndEventProcessor implements BpmnElementProcessor<ExecutableE
         // activate the compensation handler
         completedActivities.forEach(
             activity -> {
-              eventPublicationBehavior.activateCompensationHandler(activity, activated);
+              compensationSubscriptionBehaviour.activateCompensationHandler(activity, activated);
             });
       }
     }
